feat(better-output): add error catching and logging
There is a strange behavior in how HandBrakeCLI handles errors from
incorrect user commands: it doesn't set the return code to a non-zero
value.

At the same time, if an error occurs, HandBrakeCLI prints information
about it to stderr with an "ERROR" tag/prefix.

What is odd is that even when there is no error, HandBrakeCLI still
prints some service information to stderr.

So this is a little workaround to make messages as useful as it can
be. But I think it requires testing in different cases and inputs.